# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a auth server",
        "title": "LoginHub API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/passport/oauth2/provider": {
            "get": {
                "description": "列出oauth2第三方平台",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "列出oauth2第三方平台",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            },
            "post": {
                "description": "创建oauth2第三方平台",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "创建oauth2第三方平台",
                "parameters": [
                    {
                        "description": "创建oauth2第三方平台请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.CreateOauth2ProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/oauth2/provider/{provider}/delete": {
            "post": {
                "description": "删除oauth2第三方平台",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "删除oauth2第三方平台",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/oauth2/provider/{provider}/update": {
            "post": {
                "description": "更新oauth2第三方平台",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "更新oauth2第三方平台",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新oauth2第三方平台请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.UpdateOauth2ProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/oauth2/redirect/{provider}": {
            "get": {
                "description": "获取oauth2重定向url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "获取oauth2重定向url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.RespBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/passport.GetOauthRedirectURLResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.RespBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "http code",
                    "type": "integer"
                },
                "data": {
                    "description": "response data"
                },
                "msg": {
                    "description": "response message",
                    "type": "string"
                },
                "reason": {
                    "description": "reason key",
                    "type": "string"
                }
            }
        },
        "passport.CreateOauth2ProviderRequest": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret",
                "provider",
                "redirect_url"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                }
            }
        },
...
        "passport.GetOauthRedirectURLResponse": {
            "type": "object",
            "properties": {
                "redirect_url": {
                    "type": "string"
                }
            }
        },
...
        "passport.UpdateOauth2ProviderRequest": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret",
                "redirect_url"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                }
            }
        },
...
    }
}